test NodeConfig
